Extract tests for match to own file
